change to long
changed int to long

+++ src/fiboseriesrec/FiboSeriesRec.java
@@ -5,21 +5,21 @@ public class FiboSeriesRec {
     public static void main(String[] args) {
         Scanner a = new Scanner(System.in);
         System.out.println("How many numbes do you want calculated?");
-        int fi = a.nextInt();
+        long fi = a.nextLong();
         long tim = System.currentTimeMillis();
-        int b = fib(0,1,fi,true);
+        long b = fib(0,1,fi,true);
         long sTim = System.currentTimeMillis();
         long eTime = sTim - tim;
         System.out.print(eTime);
         System.out.println(" Milliseconds");
     }
-    public static int fib(int c,int d, int left,boolean bo){
+    public static int fib(long c,long d, long left,boolean bo){
         if(bo){
         System.out.println("0");
         return(fib(c,d,left-1,false));
         }else{
         if (left>0){
-        int e = c+d;
+        long e = c+d;
         System.out.println(e);
         return (fib(d,e,left-1,false));
         }
